% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{apply_IBR}
\alias{apply_IBR}
\title{Iterated Best Response solution}
\usage{
apply_IBR(game, depth = 10)
}
\arguments{
\item{game}{the signaling game that is played}

\item{depth}{depth of recursive reasoning}
}
\value{
A list with the final sender and receiver strategies.
}
\description{
Calculates the IBR solution for the signaling game
}
\examples{
apply_IBR(scalar_scalar_implicature_game)

}

