% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{apply_IQR}
\alias{apply_IQR}
\title{Iterated Quantal Response solution}
\usage{
apply_IQR(game, depth = 10, lambda = 5)
}
\arguments{
\item{game}{the signaling game that is played}

\item{depth}{depth of recursive reasoning}

\item{lambda}{"rationality"/soft-max parameter}
}
\value{
A list with the final sender and receiver strategies.
}
\description{
Calculates the IQR solution for the signaling game
}
\examples{
apply_IQR(scalar_scalar_implicature_game)

}

