% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{apply_RD}
\alias{apply_RD}
\title{Calculate expected utility for sender}
\usage{
apply_RD(game, iterations = 100, initialPertubation = 0, add = 0.01)
}
\arguments{
\item{game}{the game that is being played}

\item{rec}{receiver strategy: Message X State row-stochastic matrix}
}
\value{
A State X Message matrix of expected utilities for the sender.
}
\description{
Calculates the expected utility of a sender, given a game and the receiver's behavioral strategy.
}
\examples{
get_EU_sender(rec, scalar_implicature_game)

}

