% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{apply_RSA}
\alias{apply_RSA}
\title{Rational Speech Act solution}
\usage{
apply_RSA(game, depth = 10, lambda = 5)
}
\arguments{
\item{game}{the signaling game that is played}

\item{depth}{depth of recursive reasoning}

\item{lambda}{"rationality"/soft-max parameter}
}
\value{
A list with the final sender and receiver strategies.
}
\description{
Calculates the RSA solution for the signaling game
}
\examples{
apply_RSA(scalar_scalar_implicature_game)

}

