% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{best_response}
\alias{best_response}
\title{Calculate best response for sender}
\usage{
best_response(EU)
}
\arguments{
\item{EU}{expected utility matrix}
}
\value{
A sender strategy matrix (states times messages) with only best responses in the support.
}
\description{
Calculates the best response for a sender, given a game and the receiver's behavioral strategy.
}
\examples{
best_response(get_EU_sender(rec, scalar_implicature_game))

}

