% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.r
\name{create_game}
\alias{create_game}
\title{Create an interpretation game}
\usage{
create_game(states, messages, semantics = NULL, utils = NULL,
  prior = NULL, message_preferences = NULL)
}
\arguments{
\item{states}{character vector of state names}

\item{messages}{character vector of messages names}

\item{semantics}{a States X Messages matrix of conventional meanings}

\item{utils}{States X States matrix of utilities}

\item{prior}{vector of prior probabilities of states}

\item{message_preferences}{vector of factors of relative message preference}
}
\value{
A list representation of the interpretation game.
}
\description{
Creates a signaling game with conventionally meaningful messages
}
\examples{
show(scalar_implicature_game)

}

