% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{get_EU_receiver}
\alias{get_EU_receiver}
\title{Calculate expected utility for receiver}
\usage{
get_EU_receiver(sen, game)
}
\arguments{
\item{sen}{sender strategy: State X Message row-stochastic matrix}

\item{game}{the game that is being played}
}
\value{
A Message X State matrix of expected utilities for the receiver.
}
\description{
Calculates the expected utility of a receiver, given a game and the sender's behavioral strategy.
}
\examples{
get_EU_receiver(sen, scalar_implicature_game)

}

