% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{get_receiver_belief}
\alias{get_receiver_belief}
\title{Receiver's posterior beliefs}
\usage{
get_receiver_belief(sen, game)
}
\arguments{
\item{sen}{sender strategy: State X Message row-stochastic matrix}

\item{game}{the game that is being played}
}
\value{
A Message X States row-stochastic matrix of posterior beliefs for the receiver.
}
\description{
Calculates the posterior beliefs about the actual state of the receiver, given a game and the sender's behavioral strategy.
}
\examples{
get_receiver_belief(sen, scalar_implicature_game)

}

