% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.r
\name{get_similarity}
\alias{get_similarity}
\title{Create a utility matrix based on similarity}
\usage{
get_similarity(ns, weber_fraction = 0.2, epsilon = 0)
}
\arguments{
\item{ns}{number of states}

\item{weber_fraction}{positive real governing perceptual noise}

\item{epsilon}{baseline perceptual noise}
}
\value{
An ns X ns utility matrix
}
\description{
Creates a utility matrix loosely based on ideas from number perception.
}
\examples{
get_similarity(10)

}

