% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{apply_RD}
\alias{apply_RD}
\title{Discrete time replicator dynamics in behavioral strategies}
\usage{
apply_RD(game, iterations = 100, initialPertubation = 0, add = 0.01)
}
\arguments{
\item{game}{the signaling game that is played}

\item{iterations}{how many discrete-time update steps}

\item{add}{background fittness; influences speed of change;}

\item{initialPerturbation}{how much to wiggle the initial (semantic-play) state}
}
\value{
A list with the final sender and receiver strategies.
}
\description{
Calculates discrete time replicator dynamics in behavioral strategies for a signaling game
}
\examples{
apply_RD(scalar_scalar_implicature_game)

}

