% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_concepts.r
\name{best_response}
\alias{best_response}
\title{Calculate expected utility for sender}
\usage{
best_response(EU)
}
\arguments{
\item{rec}{receiver strategy: Message X State row-stochastic matrix}

\item{game}{the game that is being played}
}
\value{
A State X Message matrix of expected utilities for the sender.
}
\description{
Calculates the expected utility of a sender, given a game and the receiver's behavioral strategy.
}
\examples{
get_EU_sender(rec, scalar_implicature_game)

}

